add user module with route protection

+++ backend/controllers/goalsController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 /**
  * @desc Get goals
@@ -7,7 +8,7 @@ const Goal = require("../models/goalModel");
  * @access Private
  */
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json({ message: "Get Goals", data: goals });
 });
@@ -25,6 +26,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(201).json({ message: "Set Goals", data: goal });
@@ -42,6 +44,17 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found!");
+  }
+
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found!");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("You are not authorized to perform this action");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(goalId, req.body, {
@@ -66,11 +79,22 @@ const deleteGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found!");
   }
 
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found!");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("You are not authorized to perform this action");
+  }
+
   await Goal.findByIdAndDelete(goalId);
 
   res
     .status(200)
-    .json({ message: `Deleted Goal ${req.params.id}`, id : goalId });
+    .json({ message: `Deleted Goal ${req.params.id}`, id: goalId });
 });
 
 module.exports = {
